Fix visual star mappings
More recent versions of Vim seem to have changed the behavior of
functions using the expression register.  For some reason the
'normal!  gvy` command was corrupting the jump list and causing very
strange things to happen when invoked at the expression register.

Refactor the visual star mappings so that they don't need to use the
expression register and we can avoid this bug, whether intentional or
not